Create a default keypair for each node
Instead of creating a global keypair and attaching it to the images,
create the keypairs for each node, when creating them.

The previous approach caused a keypair to be created every time a
context was created (which is not good), and there was no reliable way
to identify which keypair was assigned to each node.

With this change keypairs will be created for each node with the node
name, so they can be properly managed. Now keypairs will only exists as
long as the nodes using them exist.